+++ fltk/src/app/event.rs
@@ -1,5 +1,5 @@
 use crate::app::widget::first_window;
-use crate::enums::{Event, Key, Shortcut};
+use crate::enums::{Event, Key, Shortcut, CallbackReason};
 use crate::prelude::*;
 use crate::utils::FlString;
 use fltk_sys::fl;
@@ -559,5 +559,12 @@ pub fn raw_open_callback(cb: Option<fn(*const raw::c_char)>) {
         } else {
             fl::Fl_open_callback(None)
         }
+    }
+}
+
+/// Get the callback reason
+pub fn callback_reason() -> CallbackReason {
+    unsafe {
+        mem::transmute(fl::Fl_callback_reason())
     }
 }